Begin to create better modules inside the codebase
