+++ src/attributes/index.js
@@ -1,0 +1,3 @@
+module.exports = {
+  getInitialValues: require('./initialValuesCreation')
+};
