+++ src/coercion/boolean.js
