+++ src/coercion/string.js
