--- src/errorMessages.js
@@ -1,4 +1,0 @@
-module.exports = {
-  NON_OBJECT_ATTRIBUTES: '#attributes can\'t be set to a non-object.',
-  ARRAY_OR_ITERABLE: 'Value must be iterable or array-like.'
-};
