+++ src/schema/index.js
@@ -1,0 +1,3 @@
+module.exports = {
+  normalize: require('./normalize')
+};
