+++ src/serialization/index.js
@@ -1,5 +1,5 @@
-const { SCHEMA } = require('./symbols');
-const { getType } = require('./typeResolver');
+const { SCHEMA } = require('../symbols');
+const { getType } = require('../typeResolver');
 
 function serialize(structure) {
   if(structure === undefined) {
