+++ test/unit/coercion/array.spec.js
