+++ test/unit/coercion/arraySubclass.spec.js
