+++ test/unit/coercion/boolean.spec.js
