+++ test/unit/coercion/number.spec.js
