+++ test/unit/coercion/pojo.spec.js
