+++ test/unit/coercion/string.spec.js
