+++ test/unit/coercion/structure.spec.js
