+++ test/unit/coercion/typeCoercion.spec.js
